moved class to its own folder
